requirements.txt
apache-airflow-providers-amazon==9.1.0
apache-airflow-providers-postgres==5.11.3
boto3==1.35.64
mlflow==2.17
numpy==2.0.0
pandas==2.2.2
psycopg2-binary==2.9.9
scikit-learn==1.5.1
